Do not add nullptr typeface to cache
The typeface cache is designed around deduplicating typeface ids. Since
the nullptr typeface does not have an id it makes no sense to cache it.
There is no means to get a "hit" on the nullptr typeface in any event,
since the typefaces in the cache are their own key.

This change fixes the last known user which does not check that it is
adding a nullptr typeface and adds a release assert in
SkTypefaceCache::add that the typeface is not nullptr.

This is a general follow up to "[fci] Do not add nullptr typeface to
cache" [0].

[0] https://skia-review.googlesource.com/c/skia/+/975927 .

Change-Id: I2dc253f0fee102e5bbfe4ff5e351385ba6ff88cf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/975978
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>