Remove SkScalerContext::forceGenerateImageFromPath
Instead SkScalerContext implementations should determine on a glyph by
glyph basis if the metrics should be generated from paths and, if so,
they may call SkScalerContext::generateImageFromPath to generate the
image.

SkScalerContext::fGenerateImageFromPath is made `const` and remains to
force an attempt to generate the metrics and image from paths in the
case of a stroke or path effect.

Change-Id: I32c6f5b282afdc798692d2b6f75f73bb9d8a78e5
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/975980
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>