[fontations] Flat pixel geometry only when requested
FontationsFtCompareGM::fSimulatePixelGeometry is currently only used to
set the name of the gm. Instead set the subpixel geometry to
flat/unknown only if it is set. This restores tests for known subpixel
geometry, which was regressed in "[Fontations] Match FreeType
antialiasing on unknown pixel geometry" [0]

[0] https://skia.googlesource.com/skia.git/+/13c6a67128538902d1206590fe9e256c54143e35

Change-Id: I3adeede1d1f52db575f5188b1f550c5e9dc61545
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/976196
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Dominik Röttsches <drott@google.com>