Define flag again for PNG encoding
For Bazel builds, I defined the flag under the 'png_encode' target but
not the 'png_encode_base' which is a dep of 'png_encode' which I
modified. So for these Bazel builds the flag was enabled in
SkPngEncoderImpl.cpp but not SkPngEncoderBase.cpp, which is now fixed.

Change-Id: Ia3b1d63b6e7813e0e511e4fd04df46836f763e21
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/976017
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>