Use polymorphism to clean up SkEdge
Before this change, fCurveCount was negative for cubics,
positive for quads and depended on the client to cast before
calling updateCubic or updateQuadratic respectively.

I've replaced that with a virtual function to make things
more readable (at the cost of a vtable in all SkEdges - I think
we'll manage). I've also renamed fCurveCount to fSegmentCount
and made it unsigned (since the subclass can remember what to
do now).

This change includes a few more readability things I forgot
from previous CLs.

Change-Id: I9357b70c8991aedc101ccb7b555bc643a84be372
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/976336
Reviewed-by: Daniel Dilan <danieldilan@google.com>