[graphite] Refactor VulkanRenderPass metadata in prep for AttachmentDesc update
Renames VulkanRenderPass::VulkanRenderPassMetadata to
VulkanRenderPass::Metadata.

Moves some static functions that take the metadata to just be member
functions operating on the metadata.

Removes dependent state members and switches them to function calls so
they can't get stale if the VulkanRenderPass::Metadata were to change.

Adjusts visibility of certain functions in VulkanResourceProvider and
removes friending with VulkanCommandBuffer and VulkanGraphicsPipeline so
that it's clear findOrCreateRenderPassWithKnownKey is internal to the
resource provider.

Change-Id: Icbb4f4e15600914e5a3699a343e37271ce924094
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/969102
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>