img-ddk: Fix Driver Version Deduction for Unknown ANGLE Backend
Previously, driver version checks would fail when the ANGLE backend was
unknown. This issue occurred because the version could not be correctly
determined under certain conditions.

To resolve this, the driver version is now conditionally deduced based
on the available information, ensuring that the appropriate version is
used even when the backend is not explicitly identified. This prevents
errors and improves compatibility with unknown ANGLE backends.

BUG: b/406427236
Change-Id: Ie0bc688936d0ff4217bd341d68c882cd42e9a93f
Signed-off-by: Faijur Rahman <faijur.rahman@imagination.corp-partner.google.com>
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/980376
Commit-Queue: Greg Daniel <egdaniel@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>