Fix issue where drawArc loses endcaps when stroke width is greater than or equal to oval size
In the Ganesh backend, we use 'hasStroke' instead of 'stroked' to determine whether a round cap should be drawn. This means that if the stroke width is greater than or equal to the oval width, the round cap will also be drawn.

In the Graphite backend, if the stroke width is greater than or equal to the  oval width, we need to reset 'innerRadius' and 'localInnerRadius'. This ensures that the rendering result is equivalent to drawing a filled arc plus two round caps, where the radius of the filled arc is half of the oval width plus half of the stroke width. Additionally, to correctly calculate the 'roundCapRadius', we have to introduce a new instance attribute named 'inRoundCapRadius'.

Bug: 406747427
Change-Id: I8eea35fdbe073ac08b3dcc74ad51e3fa5593e602
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/974116
Commit-Queue: Jim Van Verth <jvanverth@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>