Allow mask filters to effect shading properties for autolayer
In order to implement embossMaskFilter, we needed a way to have the
output of the maskfilter affect the shading of a layer, before the image
filter, in order to replicate current behavior. To do this we have the
the maskfilter asImageFilter() function return whether it is just the
coverage or if it will return an rgba image after the filter. When
restoring the layer, this will draw that result into the dst [1].

The new impl for asImageFilter() takes in a paint that it uses to
determine the original color/shading properties. It can then apply
filters to it and mask it by the original or a different coverage.

Although embossMaskFilter is technically deprecated, this is necessary
because some clients still rely on the raster version of it, and we
don't want to break them when we deprecate the old mask filter
implementations.

Due to the fact that we are now blending the coverage to the shading in
the mask filter, results may look incorrect for draws with certain blend
modes for emboss. This is a tradeoff we are willing to make due to the
fact that this function is technically deprecated, as well as in java
for android. This will look correct for SrcOver blends, which is the
only usages for this mask filter in Android Apps as far as we can tell.

Two more caveats to this implementation:
- In order to replicate the original effect we extended the current lighting shader. This is still going to be slightly different than the original due to coverage blur calculation, normal vector calculation, and precision being different.
    - SkImageFilters.h
    - SKLightingImageFilter.cpp
    - sksl_rt_shader.sksl
    - sksl_rt_shader.minified.sksl
    - sksl_rt_shader.unoptimized.sksl
- Small text with this mask filter will be drawn as a path so that the autolayer is applied and gets the mask filter effect, otherwise we don't currently have a way to apply it internally and cache it yet, but will be done in the future.
    - SubRunContainer.cpp

[1]
https://source.chromium.org/chromium/chromium/src/+/main:third_party/skia/src/core/SkCanvas.cpp;drc=26fa5b6a22b36be60bc17cccd5754c44dcff3334;l=872

Bug: 304823483
Change-Id: Ie41818a630e7fd4ecb66fca5e2afa67a9a5f1195
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/978432
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>