[graphite] Clamp quantized mask blur values to 1/32
Updates `quantize` to return at minimum a value of 1/32 so that small
positive values aren't rounded to 0, which then caused infinities to
be generated later on, leading to eventual overflows.

Updates the circle profile generation to also check for overflows,
and adds a unit test reproducing the crash. Confirmed that even w/o
the quantize fix, the circle profile overflow guarding works and would
cause Graphite to just fallback to the image-based blur path.

Bug: b/413427423
Change-Id: I84c3fe0c71ed31b59d8b752b1517ac141b5e7172
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/986437
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>