[graphite][vulkan] Use extended dynamic state
This change turns all the state that Graphite always sets to
a constant value dynamic, which get set once at the beginning of the
render pass.

It also makes the topology and practically all of depth/stencil state
dynamic (assuming driver support and no bug workarounds).

As it is implemented entirely in the backend, it is not quite as
efficient as it can be.  The front-end still sees the pipelines as
different objects; but now creation of a pipeline whose only different
state is depth/stencil would hit the cache if it previously didn't.

Another source of inefficiency is that on pipeline bind, the state needs
to be diffed with the previous pipeline to know what's changed.  The
front-end might be able to do better.

Bug: skia:414645289
Change-Id: Ie61f6aec8626c5161266fb90bec3c48c530f2c91
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/989157
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>