Make sure inline static symbols in headers are visible
Objects which are defined in headers but are supposed to be globally
unique (e.g. static members) may accidentally get duplicated if the
header is built into multiple shared libraries. This can cause very
subtle bugs. This CL prevents that by marking them as visible, so the
linker knows to deduplicate the copies.

Bug: chromium:368047728
Change-Id: I148c2e07095e839875e21507092f1020adcc546c
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/988456
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Ben Wagner <bungeman@google.com>