[graphite][vulkan] Use VK_EXT_graphics_pipeline_library
This extension can be used to create subsets of a pipeline separately,
then link them.  In particular, this forces many vendors that would
otherwise bake the blend state in the fragment shader (and cause
pipeline explosition) to instead patch a compiled fragment shader with
blend code.  For these vendors (where blend info is not dynamic state),
this is what they would do in their GL drivers.

In this change, use of this extension is implemented entirely in the
Vulkan backend by first creating the "shaders" subset of the pipeline,
and then creating the full pipeline (with the vertex input and fragment
output state) immediately afterwards using the shaders library, which is
a cheap operation.

This relies on the shaders library to hit the cache and is not the most
optimal usage of the extension.  Ideally, the front-end creates fewer
shader library pipelines and creates multiple full pipelines out of them
as needed.

Bug: skia:414645289
Change-Id: I2960ef42a7af1617f0359013d10b018663fdfe89
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/989296
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>