[pdf] Must add all reserved references
Once an object reference is reserved, exactly one object must be emitted
with that reference. This was not well documented, so document it.

SkPDFStructTree::makeOutline was violating this requirement when a
structure tree was provided and a header outline was requested but no
header content was found (in which case there isn't really any outline).
Fix this by creating the header outline structure and reserve object
references afterword only if the structure is about to be emitted. Add a
test that failed (asserted or produced an invalid PDF) before this
change.

Change-Id: Idbb79f0831fa47dac8a7b35e7830a90147a90942
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/989699
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>