[graphite][vulkan] Minimize graphics pipeline key
This change removes unnecessary bits from the render pass description
when creating the graphics pipeline key.  In particular, things that
keep the render pass compatible, such as load/store op do not need to
contribute to the graphics pipeline key.  Additionally, the existence of
a resolve attachment and whether load-from-resolve happens can all be
encoded with one bit.

This brings the graphics pipeline key size and complexity in line with
the other backends.

Bug: skia:414645289
Change-Id: I32af49b19c9b9c0f19044c24ca2a41b45c6322c2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/990096
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>