GraphiteDawn: support rendering with smaller MSAA texture.
In many scenarios, only a portion of the target texture needs to be
updated. In those cases, if supported, we can use a smaller MSAA texture
to draw, then resolving to the target texture using non-zero offset.

For this to work, we need to offset the render pass' viewport such that
the draw bounds' top left corner will be at (0, 0) on the MSAA texture.

At the end the render pass, we will resolve from:
- The MSAA texture's (0, 0, bounds.width, bounds.height) area
to:
- The target texture's (bounds.x, bounds.y, bounds.width, bounds.height)
area.

Bug: b/415842843
Change-Id: I13017706f8c1f212d9292d72ceb98276dd625a66
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/990976
Commit-Queue: Quyen Le <lehoangquyen@chromium.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>