[graphite][vulkan] Make render pass handling less special
Instead of processing RenderPassDesc into VulkanRenderPass::Metadata,
this change makes the backend directly use RenderPassDesc everywhere.
The Metadata class's preprocessing of the RenderPassDesc was really
minimal, and directly processing RenderPassDesc would be more common
when support for dynamic rendering is added.

The downside of the Metadata class was that it mode the render pass seem
more generic than it really was.  RenderPassDesc on the other hand comes
with invariants that can simplify the code; this is relied upon in the
upcoming change that adds support for
VK_EXT_multisampled_render_to_single_sampled.
As a side effect of those invariants, the render pass key is minimized
to one uint32 in this CL.

Bug: skia:415773542
Change-Id: I7c6a560ab0ae8f9b6c85a3fb52656e571ff8bcd2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/991418
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>