[graphite] Fix device-space blur mask filters with non-uniform scale
Blur mask filters are created with a uniform blur radius. In Graphite's
improved mask filter model, this can be transformed into a non-uniform
blur if the mask filter respects the CTM and the CTM has different
scale factors per axis.

On the other hand, if the mask filter ignores the CTM, it should always
be a uniform blur in device space. The way the mapping was implemented
to bring it back to a "local" blur for the image filter assumed that
the CTM had a uniform scale. In this situation, the resulting local
blur's uniform scale would be transformed to a non-uniform device
space blur.

Bug: chromium:411304524
Change-Id: I2b77ae0af0f3c38badb7f0e4ed826b55725c1329
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/991136
Reviewed-by: Daniel Dilan <danieldilan@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>