[graphite][vulkan] Simplify render pass creation
If the color attachment's initial and final layouts are
COLOR_ATTACHMENT_OPTIMAL, it's known through experience that using
the GENERAL layout for that attachment in a subpass is optimal.

This change simplifies render pass creation by removing the need for
special handling render passes that use or not an input attachment; the
render passes are always created with the assumption that input
attachments may be used.  This, has no performance cost, and removes
some complications with dealing with render pass compatibility.

This change is in preparation for removing VulkanRenderPass::Metadata
and using RenderPassDesc directly.

Bug: skia:415773542
Change-Id: I72fbf9d900444e3eb39cbbc857f0b4336c904c88
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/991417
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Auto-Submit: Shahbaz Youssefi <syoussefi@google.com>
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>