[graphite] Simplify LocalMatrix shader expression w/ and w/o perspective
For non-perspective, it switches from 4x4 to a 2x2 packed as a float4
and a separate float2. While most compilers probably could reduce the
4x4 matrix multiply to an equivalent number of operations given the
hardcoded z=0,w=1, this reduces the amount of data per block from 16 to
6.

For perspective, it switches from 4x4 to 3x3 for a data reduction of 16
to 9.

With LM blocks being able to be lifted to the vertex shader, this is
less likely to make a performance boost, and I didn't see any GPU frame
time improvements on my M1. If nothing is detected on perf, we can
choose to revert this.

Change-Id: Idea95d0c92f96972c93f9c560d81911fd82a41e7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/991396
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>