[graphite][vulkan] Simplify render pass key
This is just the CL https://skia-review.googlesource.com/c/skia/+/989997 pulled out of its CL stack.
I need this, sooner rather than later, for the Android/Vulkan Precompile CL.

The Vulkan render pass key was quite generic, but in practice there is
only ever a need for single-subpass render passes.  That is except for
emulating VK_EXT_multisampled_render_to_single_sampled where a very
specific unresolve subpass is added.

This change simplifies the key based on the above assumptions.  It
also includes an implementation for VulkanCaps::extractGraphicsDescs
which converts the pipeline key (including the render pass key) back to
the pipeline/render pass description.

A future change can further simplify the pipeline key in particular
based on render pass compatibility rules, so that the key is reduced to
4 uint32s like the other backends.

With this change, the PaintParamsKeyTest$ DM test no longer fails on
assertion with the grvk config.

Bug: skia:414645289
Bug: b/390186667
Change-Id: I79b731cd2495e8dc4d6892a3e029992cfec9e74f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/993543
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>