[sksl] Skip inline statement limit for modules
This was added as an upper bound for pathologic inputs from user
provided SkSL. However, it turns out that the Graphite modules have
gotten large enough that their base level of statements quickly pushes
inlining past the 2500 statement limit.

This would mean that somewhat arbitrarily simple seeming module
functions would remain un-inlined within the modules. These would not
be re-inlined when the final program was compiled because that would
only execute a single iteration of inlining. This would be enough to
inline the primary module function calls, but not enough to then
recurse and inline things that were missed when compiling as a large
module stopped the process early.

I'll be curious to see how this impacts performance, but I couldn't
measure improvements on Mac. That's not surprising given they use
llvm on the generated msl so it can clean these issues up. That said
I think always maximally inlining the module functions produces
more predictable output.

NOTE: This does not actually impact the .unoptimized.sksl and
.minified.sksl because inlining is executed as part of loading
the modules.
Change-Id: I8739dd5d38649161c00072d2f1bf1494c106ab21
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/993544
Reviewed-by: Robert Phillips <robertphillips@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>