[graphite] Fix for Clip key collision
Changes the bounds used in the key or the subset search to be
relative to the entire transformed mask. Otherwise it can be
ambiguous if the device bounds and mask are translated relative
to each other.

Also includes a number of name changes to make things clearer.

Bug: b/412707592
Change-Id: Id581371d512b1205a536202b636389542f218516
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/991726
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>