Implement drawCoverageMask for Ganesh
The matrix that drawCoverageMask receives places the coverage mask in
device space, but shading is done in the local space. Additionally, the
devices current localToDevice matrix may have perspective transforms
that maskToDevice doesn't. Thus, this function uses maskToDevice and
localToDevice to transform the coverage mask to local space for drawing.

This change currently isn't load bearing until we set
useDrawCoverageMaskForMaskFilters for ganesh/Device to true, which will
happen when the fast path drawblurredrrect and caching on the canvas
level is implemented.

When enabled, this changes the behavior of some mask filters (bounds of
tablemaskfilter and enables embossmaskfilter to have an implementation).

Bug: 304823483
Change-Id: I508714e8819dce856290b5405b0794844df876b7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/990556
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>