Add initial implementation of SkRasterContext and SkRasterRecorder
There are a few objectives here:
 - Allow a better way to capture multiple instructions across multiple surfaces to allow us to have a better debugging and performance analysis story with big clients.
 - Reduce hard-coded global caches (configured at compile time)
 - Align our public APIs and behavior between Raster and Graphite (and Ganesh, sort of)
 - Enable frame-to-frame performance optimizations

The first step is to mirror Graphite's concept of a Context (created
once) [1] and Recorders [2] (created potentially multiple times).
A surface is created from a recorder. When a draw call is made to
that surface's canvas, it will get passed to the recorder and
(eventually) sent to the context (e.g. "flushed"). Raster surfaces
are currently immediate and synchronous, but we could have a
SkRasterImmediateRecorder and SkRasterAsyncRecorder one day to have
both options available.

This CL introduces a `SkRasterContext` and `SkRasterRecorder`, both
of which are intended to be opaque, with their actual implementations
tucked away in `src/core/SkRaster(Context|Recorder)_Base`. This is
plumbed into some places in the Raster backend and future CLs will
add functionality to it.

Suggested review order:
 - tests/RasterContextTest.cpp to get a sense of the new APIs
   as they would be called by clients (see how Graphite contexts
   and Ganesh contexts can make a raster recorder still, so clients
   only need the one type of context)
 - include/core/SkRaster*
 - Other changes to C++ files in include/core (e.g. Graphite)
 - src/core/SkRaster*
 - SkSurface_Raster.*, SkBitmapDevice.*, SkDrawBase.* to see how
   SkRasterRecorder (and eventually some of the things on it) will
   be piped through.
 - All other changes in any order.

[1] https://github.com/google/skia/blob/2dbbb5680507ff00effa98e9e73dcb4dde5bfde3/include/gpu/graphite/Context.h
[2] https://github.com/google/skia/blob/2dbbb5680507ff00effa98e9e73dcb4dde5bfde3/include/gpu/graphite/Recorder.h

Change-Id: Ie19f1d836e47775c07778b872c117869718c9166
Bug: b/385170812
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/977817
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>