[graphite] Add two Android-specific RuntimeEffect-based Precompile Options
This CL adds support for the Pipelines using:
    RE_MouriMap_BlurEffect
and RE_MouriMap_TonemapEffect.

The hope is that we can expand Android's MouriMap class a bit (adding the names and the SkRuntimeEffect accessors) and then just reuse it for Precompile.

This CL also fixes a bug where Precompilation wasn't adding the KeyBlocks for the toLinearSrgb/fromLinearSrgb intrinsics.

Bug: b/390186667
Change-Id: I33439b7d99631a857fd891eca0a49b75a7a3b724
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/996116
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>