[Fontations] Match FreeType backend for autohinting selection on Android
On Android, the FreeType backend only enables autohinting when the force flag is set and prevents fallback otherwise regardless of the requested hinting mode or presence of native hints in the font. This changes the Fontations backend to match that behavior.

This is an imported pull request from
https://github.com/google/skia/pull/199

GitOrigin-RevId: 2613b7852d8f2cf140c383c7067af76e14c57082
Change-Id: If75e0a3f417f90392e2655acc1fa6cb20c990360
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/997516
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>